Added some small scripts
Added some scripts. Some of them are kinda useless but I usually try to wrap basic functions in other scripts in case I want to extend the functionality. E.g., I have my set-brightness.sh script, which currently only runs a basic brightnessctl command, but if I want to, idk, check if my brightness is too low or too high, I can prevent it from being too low or too high. Or maybe run my get-battery.sh script in order to check if my battery is too low, so that I can artificially limit my maximum brightness, as a battery-saving measure? Not sure if this wrapping is a common thing and it might not be the most elegant, but I think it's somewhat elegant, and it's super convenient for me for scripting purposes. Like every function I have just has an appropriate script with it, so I can just edit that script and anytime I want to set my brightness or fetch my battery level, I can just call this one script.